delete old pro page
