+++ packages/app/src/app/pages/Pro/index.tsx
@@ -1,278 +1,121 @@
-import { format } from 'date-fns';
-import { Helmet } from 'react-helmet';
 import React from 'react';
-import { useLocation } from 'react-router-dom';
-
+import { Helmet } from 'react-helmet';
+import { Switch, Route } from 'react-router-dom';
+import { ThemeProvider, Stack } from '@codesandbox/components';
+import css from '@styled-system/css';
 import { useOvermind } from 'app/overmind';
+import { Step } from 'app/overmind/namespaces/pro/types';
 import { Navigation } from 'app/pages/common/Navigation';
-import MaxWidth from '@codesandbox/common/lib/components/flex/MaxWidth';
-import Margin from '@codesandbox/common/lib/components/spacing/Margin';
-import Centered from '@codesandbox/common/lib/components/flex/Centered';
+import { NewTeam } from 'app/pages/common/NewTeam';
 
-import { ThemeProvider, Switch } from '@codesandbox/components';
-
-import { SubscribeForm } from './subscribe-form';
-import {
-  Page,
-  Content,
-  Avatar,
-  Badge,
-  Button,
-  ButtonAsLink,
-  Heading,
-  HelpText,
-  LinkButton,
-  SubHeading,
-  DurationChoice,
-  BillText,
-} from './elements';
+import { WorkspacePlanSelection } from './pages/WorkspacePlanSelection';
+import { InlineCheckout } from './pages/InlineCheckout';
+import { ConfirmBillingInterval } from './pages/ConfirmBillingInterval';
+import { PaymentSuccess } from './pages/PaymentSuccess';
 import { SignInModalElement } from '../SignIn/Modal';
-import { getUserExpiringDate } from './getExpiringDate';
-import { ProPage as ProPageV2 } from './index-v2';
 
 const ProPage: React.FC = () => {
   const {
-    state: { hasLoadedApp, isLoggedIn, user, patron },
+    state: { hasLoadedApp, isLoggedIn },
     actions: {
-      modalOpened,
-      patron: {
-        priceChanged,
-        createSubscriptionClicked,
-        cancelSubscriptionClicked,
-        updateSubscriptionClicked,
-        patronMounted,
-      },
+      pro: { pageMounted },
     },
   } = useOvermind();
 
-  const checkoutDisabled = !hasLoadedApp || !isLoggedIn;
-
-  // silly hack to allow cached versions to keep working
-  priceChanged({ price: 12 });
-
   React.useEffect(() => {
-    patronMounted();
-  }, [patronMounted]);
+    pageMounted();
+  }, [pageMounted]);
 
-  const location = useLocation();
-
-  const getContent = () => {
-    /**
-     * Proceed with caution
-     * There is logic baked in the order of
-     * the conditions.
-     */
-    if (!hasLoadedApp) return null;
-
-    if (!isLoggedIn) return <SignInModalElement />;
-
-    if (!user.subscription) {
-      return (
-        <NotPro
-          createSubscriptionClicked={createSubscriptionClicked}
-          user={user}
-          patron={patron}
-          checkoutDisabled={checkoutDisabled}
-        />
-      );
-    }
-
-    if (user.subscription.plan === 'patron') return <Patron user={user} />;
-
-    /** Pro subscriptions land ⬇️ */
-
-    if (user.subscription.cancelAtPeriodEnd) {
-      return (
-        <Expiring
-          user={user}
-          updateSubscriptionClicked={updateSubscriptionClicked}
-          isUpdatingSubscription={patron.isUpdatingSubscription}
-        />
-      );
-    }
-
-    if (user.subscription.plan === 'pro') {
-      return (
-        <Pro
-          user={user}
-          modalOpened={modalOpened}
-          cancelSubscriptionClicked={cancelSubscriptionClicked}
-        />
-      );
-    }
-
-    return null;
-  };
-
-  if (location.search.includes('v=2')) return <ProPageV2 />;
+  if (!hasLoadedApp) return null;
+  if (!isLoggedIn) return <SignInModalElement />;
 
   return (
     <ThemeProvider>
-      <Page>
+      <Helmet>
+        <title>Pro - CodeSandbox</title>
+        <link
+          href="https://fonts.googleapis.com/css?family=Lato:400,700&display=swap"
+          rel="stylesheet"
+        />
+      </Helmet>
+      <Stack
+        direction="vertical"
+        css={css({
+          backgroundColor: 'grays.900',
+          color: 'white',
+          width: '100vw',
+          minHeight: '100vh',
+          fontFamily: 'Lato, sans-serif',
+        })}
+      >
         <Navigation title="CodeSandbox Pro" />
 
-        <Helmet>
-          <title>Pro - CodeSandbox</title>
-        </Helmet>
-
-        <Margin vertical={1.5} horizontal={1.5}>
-          <MaxWidth width={1024}>
-            <Content>{getContent()}</Content>
-          </MaxWidth>
-        </Margin>
-      </Page>
+        <Switch>
+          <Route path={`/pro/create-workspace`}>
+            <NewTeam redirectTo="/pro" />;
+          </Route>
+          <Route path={'/pro/success'}>
+            <PaymentSuccess />
+          </Route>
+          <Route path={`/pro`}>
+            <StartOrModifySubscription />
+          </Route>
+        </Switch>
+      </Stack>
     </ThemeProvider>
   );
 };
 
-const Pro = ({ user, modalOpened, cancelSubscriptionClicked }) => {
-  const subscriptionDate = new Date(user.subscription.since);
-  return (
-    <MaxWidth width={400}>
-      <Centered horizontal>
-        <Avatar src={user.avatarUrl} />
-        <Badge type="pro">Pro</Badge>
-        <Heading>You&apos;re a Pro!</Heading>
+const StartOrModifySubscription = () => {
+  /**
+   * The user flow forks based on the workspace
+   *
+   * step 1 - choose workspace and plan
+   * step 2 -
+   *    if subscription doens't exist - Inline Checkout
+   *    if subscription already exists - Confirm Amount Change
+   * step 3 - redirect to pro/success
+   *
+   * We don't treat these as seperate pages because steps transition into
+   * each other with data from the previous step
+   */
 
-        <ButtonAsLink href="/s/" style={{ marginTop: 30 }}>
-          Create a sandbox
-        </ButtonAsLink>
-
-        <HelpText>
-          You will be billed{' '}
-          {user.subscription.duration === 'yearly' ? (
-            <>
-              and charged{' '}
-              <b>annually on {format(subscriptionDate, 'MMM dd')}</b>
-            </>
-          ) : (
-            <>
-              on the <b>{format(subscriptionDate, 'do')} of each month</b>
-            </>
-          )}
-          . You can{' '}
-          <LinkButton
-            onClick={e => {
-              e.preventDefault();
-              modalOpened({
-                modal: 'preferences',
-                itemId: 'paymentInfo',
-              });
-            }}
-          >
-            update your payment details
-          </LinkButton>{' '}
-          or{' '}
-          <LinkButton
-            onClick={e => {
-              e.preventDefault();
-              cancelSubscriptionClicked();
-            }}
-          >
-            cancel your subscription
-          </LinkButton>{' '}
-          at any time.
-        </HelpText>
-      </Centered>
-    </MaxWidth>
-  );
-};
-
-const Patron = ({ user }) => (
-  <MaxWidth width={400}>
-    <Centered horizontal>
-      <Avatar src={user.avatarUrl} />
-      <Badge type="patron">Patron</Badge>
-      <Heading>You&apos;re a Pro!</Heading>
-
-      <SubHeading>
-        Thank you for being an early supporter of CodeSandbox. As a patron, you
-        can access all Pro features.
-      </SubHeading>
-      <ButtonAsLink href="/s/">Create a sandbox</ButtonAsLink>
-
-      <HelpText>
-        You will be billed on the{' '}
-        <b>{format(new Date(user.subscription.since), 'do')}</b> of each month.
-        You can <a href="/patron">modify your contribution</a> at any time.
-      </HelpText>
-    </Centered>
-  </MaxWidth>
-);
-
-const NotPro = ({
-  createSubscriptionClicked,
-  user,
-  patron,
-  checkoutDisabled,
-}) => {
-  const [duration, setDuration] = React.useState('yearly');
+  const {
+    state: {
+      pro: { step, isPaddleInitialised, isBillingAmountLoaded },
+    },
+  } = useOvermind();
 
   return (
-    <>
-      <Heading>CodeSandbox Pro</Heading>
-      <SubHeading>
-        {duration === 'yearly' ? '$9/month billed annually' : '12$/month'}
-      </SubHeading>
-      <DurationChoice>
-        <BillText on={duration === 'monthly'}>Bill monthly</BillText>
-        <Switch
-          onChange={() =>
-            setDuration(d => (d === 'yearly' ? 'monthly' : 'yearly'))
+    <Stack
+      justify="center"
+      align="center"
+      css={css({ fontSize: 3, width: 560, marginTop: 120, marginX: 'auto' })}
+    >
+      {step === Step.WorkspacePlanSelection ||
+      (step === Step.InlineCheckout && !isPaddleInitialised) ||
+      (step === Step.ConfirmBillingInterval && !isBillingAmountLoaded) ? (
+        <WorkspacePlanSelection
+          loading={
+            (step === Step.InlineCheckout && !isPaddleInitialised) ||
+            (step === Step.ConfirmBillingInterval && !isBillingAmountLoaded)
           }
-          on={duration === 'yearly'}
         />
-        <BillText on={duration === 'yearly'}>Bill annually</BillText>
-      </DurationChoice>
-      <Centered horizontal>
-        <SubscribeForm
-          subscribe={({ token, coupon }) =>
-            createSubscriptionClicked({ token, coupon, duration })
-          }
-          isLoading={patron.isUpdatingSubscription}
-          hasCoupon
-          name={user && user.name}
-          error={patron.error}
-          disabled={checkoutDisabled}
-        />
-      </Centered>
-    </>
-  );
-};
-
-const Expiring = ({
-  user,
-  updateSubscriptionClicked,
-  isUpdatingSubscription,
-}) => (
-  <MaxWidth width={500}>
-    <Centered horizontal>
-      <Avatar src={user.avatarUrl} />
-      <Badge type="pro" style={{ opacity: 0.8 }}>
-        Pro
-      </Badge>
-      <Heading>Your subscription is expiring</Heading>
-
-      <HelpText>
-        Your subscription will be automatically cancelled on your next billing
-        date ({getUserExpiringDate(user.subscription)}). All your private
-        sandboxes will remain available and private.
-      </HelpText>
-
-      {isUpdatingSubscription ? (
-        <Button disabled style={{ marginTop: 30 }}>
-          Creating subscription...
-        </Button>
-      ) : (
-        <Button
-          onClick={() => updateSubscriptionClicked({ coupon: '' })}
-          style={{ marginTop: 30 }}
+      ) : null}
+      {step === Step.InlineCheckout && (
+        <div
+          style={{
+            width: isPaddleInitialised ? 'auto' : 0,
+            height: isPaddleInitialised ? 'auto' : 0,
+            overflow: 'hidden',
+          }}
         >
-          Reactivate subscription
-        </Button>
+          <InlineCheckout />
+        </div>
       )}
-    </Centered>
-  </MaxWidth>
-);
+      {step === Step.ConfirmBillingInterval && <ConfirmBillingInterval />}
+    </Stack>
+  );
+};
 
 export default ProPage;
